feat(landing): living shared mascot + Lucide icons on the apex chooser
Replace the chooser's static mascot image and emoji button icons with the
same animated mascot the chat uses and inline Lucide SVGs, factoring the
mascot into a component shared by both surfaces.

- site/mascot.css (new): single source of truth for the mascot compositing
  rules (still + video + eyes + per-pose pupil placement + .brand-logo
  sizing). Moved out of gizza.css.
- site/mascot.js (new): ES module exporting initMascot(rootEl, opts). Holds
  the full still/video/eyes state machine + cursor-tracking pupils, extracted
  verbatim from gizza-app.js. `{ landing: true }` runs idle-only (eyes follow
  + idle video, no chat-typing states) for the chooser.
- site/gizza-app.js: init the mascot via initMascot() in chat mode and keep
  the exact `window.__brandStartTyping` / `window.__brandStopTyping` globals
  the composer-submit path drives. Chat behavior unchanged.
- site/gizza.css: drop the mascot rules (now in mascot.css); keep `.brand h1`.
- src/blocks/ui.rs: link /mascot.css in render_chat's head (mascot CSS moved
  out of gizza.css); mascot markup unchanged.
- site/index.html: living .brand-mascot markup + /mascot.css + a module that
  imports /mascot.js in landing mode; two emoji spans replaced with inline
  Lucide message-circle (Chat) and wrench (Tools) SVGs (currentColor → orange
  on hover). Page stays inert: no SW, no runtime; colors + WebGPU probe kept.
- solobase.toml: overlay site/mascot.{js,css} → /mascot.{js,css} and append
  both to extra_bypass_prefix so the SW never intercepts them.
